add Post

+++ src/Book.Service/Dtos.cs
@@ -2,8 +2,8 @@ namespace Book.Service.Dtos {
 
     public record BooksDto(int Id, string BookName, string Author, double Price, DateTimeOffset ReleaseDate);
 
-    public record CreateBookDto(string BookName, string Author, decimal Price);
+    public record CreateBookDto(string BookName, string Author, double Price);
   
-    public record UpdateBookDto(string BookName, string Author, decimal Price);
+    public record UpdateBookDto(string BookName, string Author, double Price);
 
 }
